# openapi/vgselect-3a.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VG Select: 3A (Automated Agentic Architecture)",
    "summary": "Scans a repository and workload profile, then recommends an agentic application architecture that balances latency and accuracy.",
    "description": "VG Select: 3A ranks ten agent topologies (single call, single agent, prompt chain, router, parallel sectioning, voting, evaluator-optimizer, orchestrator + workers, hierarchical, handoffs) using evidence-cited rules, estimates latency and cost for each, and produces a concrete orchestrator/subagent decomposition plan.\n\nTypical flow: `POST /api/v1/scan` (or `/api/v1/scan/upload`) to infer a profile from code, review the inferred fields, then `POST /api/v1/recommend` with corrections.",
    "contact": {
      "name": "VG Select: 3A"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/catalog/validate": {
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Validate catalog entries (ModelSpec objects) without storing them",
        "operationId": "api_catalog_validate_api_v1_catalog_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array",
                "title": "Models"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Api Catalog Validate Api V1 Catalog Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/recommend": {
      "post": {
        "tags": [
          "recommend"
        ],
        "summary": "Recommend a topology for a profile",
        "operationId": "api_recommend_api_v1_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommend/pdf": {
      "post": {
        "tags": [
          "deliverables"
        ],
        "summary": "Architecture document (PDF)",
        "operationId": "api_recommend_pdf_api_v1_recommend_pdf_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeliverableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "PDF architecture document",
            "content": {
              "application/pdf": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommend/skeleton": {
      "post": {
        "tags": [
          "deliverables"
        ],
        "summary": "Downloadable agent skeleton (zip) for the recommended topology",
        "operationId": "api_recommend_skeleton_api_v1_recommend_skeleton_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeliverableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Zip archive of a runnable project skeleton",
            "content": {
              "application/zip": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scan": {
      "post": {
        "tags": [
          "scan"
        ],
        "summary": "Scan a repository by server path or git URL",
        "operationId": "api_scan_api_v1_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scan/upload": {
      "post": {
        "tags": [
          "scan"
        ],
        "summary": "Scan an uploaded zip archive of a repository",
        "operationId": "api_scan_upload_api_v1_scan_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_scan_upload_api_v1_scan_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/describe": {
      "post": {
        "tags": [
          "recommend"
        ],
        "summary": "Infer a profile from prose with Claude, then recommend",
        "operationId": "api_describe_api_v1_describe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DescribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DeliverableRequest": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/WorkloadProfileIn",
            "description": "Workload profile. Omitted fields take engine defaults."
          },
          "include_markdown": {
            "type": "boolean",
            "title": "Include Markdown",
            "description": "Include the full Markdown report in the response.",
            "default": true
          },
          "policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SelectionPolicyIn"
              },
              {
                "type": "null"
              }
            ],
            "description": "Model selection policy; default policy uses the whole catalog, verified and approved entries only."
          },
          "catalog_models": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Models",
            "description": "Extra or overriding model catalog entries (ModelSpec objects) merged over the server's catalog for this request."
          },
          "scan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scan",
            "description": "The `scan` object returned by a previous /api/v1/scan call, to include scan findings in the document."
          },
          "framework": {
            "type": "string",
            "const": "langgraph",
            "title": "Framework",
            "description": "Skeleton framework (skeleton endpoint only).",
            "default": "langgraph"
          }
        },
        "type": "object",
        "required": [
          "profile"
        ],
        "title": "DeliverableRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecommendRequest": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/WorkloadProfileIn",
            "description": "Workload profile. Omitted fields take engine defaults."
          },
          "include_markdown": {
            "type": "boolean",
            "title": "Include Markdown",
            "description": "Include the full Markdown report in the response.",
            "default": true
          },
          "policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SelectionPolicyIn"
              },
              {
                "type": "null"
              }
            ],
            "description": "Model selection policy; default policy uses the whole catalog, verified and approved entries only."
          },
          "catalog_models": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Models",
            "description": "Extra or overriding model catalog entries (ModelSpec objects) merged over the server's catalog for this request."
          }
        },
        "type": "object",
        "required": [
          "profile"
        ],
        "title": "RecommendRequest"
      },
      "RecommendationOut": {
        "properties": {
          "product": {
            "type": "string",
            "title": "Product"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "profile": {
            "additionalProperties": true,
            "type": "object",
            "title": "Profile"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "primary": {
            "type": "string",
            "title": "Primary"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateOut"
            },
            "type": "array",
            "title": "Candidates"
          },
          "frontier": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Frontier"
          },
          "fastest_viable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fastest Viable"
          },
          "most_accurate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Most Accurate"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanOut"
          },
          "mermaid": {
            "type": "string",
            "title": "Mermaid"
          },
          "svg": {
            "type": "string",
            "title": "Svg",
            "description": "Standalone SVG rendering of the plan (no external dependencies).",
            "default": ""
          },
          "scan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scan"
          },
          "selection": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selection",
            "description": "Model selection per role: requirements, chosen model, effort, fallback, alternatives, rejections, warnings."
          },
          "selected_estimate": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Estimate",
            "description": "Topology estimate re-run on the chosen models."
          },
          "tiers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tiers",
            "description": "Reference model per capability level used for the ranking estimates."
          },
          "report_markdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Markdown"
          }
        },
        "type": "object",
        "required": [
          "product",
          "version",
          "profile",
          "headline",
          "primary",
          "candidates",
          "frontier",
          "fastest_viable",
          "most_accurate",
          "plan",
          "mermaid"
        ],
        "title": "RecommendationOut"
      },
      "ScanOut": {
        "properties": {
          "scan": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scan"
          },
          "profile": {
            "additionalProperties": true,
            "type": "object",
            "title": "Profile",
            "description": "Merged profile (scan inferences + overrides) that was recommended on."
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendationOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "scan",
          "profile"
        ],
        "title": "ScanOut"
      },
      "ScanRequest": {
        "properties": {
          "policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SelectionPolicyIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "catalog_models": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Models"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path",
            "description": "Server-side directory to scan. Must be under one of VGSELECT_SCAN_ROOTS."
          },
          "git_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Url",
            "description": "Git URL to clone (shallow) and scan. Requires VGSELECT_ALLOW_GIT_CLONE=1."
          },
          "overrides": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkloadProfileIn"
              },
              {
                "type": "null"
              }
            ],
            "description": "Profile fields that override scan inferences."
          },
          "recommend": {
            "type": "boolean",
            "title": "Recommend",
            "description": "Also run the recommender on the merged profile.",
            "default": true
          },
          "include_markdown": {
            "type": "boolean",
            "title": "Include Markdown",
            "default": true
          }
        },
        "type": "object",
        "title": "ScanRequest"
      },
...
    }
  },
...
}